+++ app/mailers/alchemy/messages_mailer.rb
@@ -1,5 +1,5 @@
 module Alchemy
-  class MessagesMailer < ActionMailer::Base
+  class MessagesMailer < BaseMailer
     def contact_form_mail(message, mail_to, mail_from, subject)
       @message = message
       mail(