+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## 3.4.0 (unreleased)
+
+__New Features__
+
+* `MessagesMailer` (formerly known as `Messages`) now inherits from `ApplicationMailer`
+when it is defined.
+
+__Notable Changes__
+
+* `Messages` mailer class has been renamed to `MessagesMailer`
+
 ## 3.3.0 (2016-05-18)
 
 __New Features__
@@ -46,7 +57,6 @@ __Notable Changes__
 * Resource CSV export now includes ID column and does not truncate large text columns anymore
 * `Alchemy::Attachment#urlname` now returns always an escaped urlname w/o format suffix and does not convert the `file_name` once on create anymore
 * Speed up the admin interface significantly when handling a large amount of pages
-* `Messages` mailer class has been renamed to `MessagesMailer`
 
 __Fixed Bugs__